requirements.txt
Selenium
python-dotenv